refactor: extract Painel tab into dedicated screen
